+++ src/formatters/functions.rs
@@ -97,6 +97,13 @@ pub fn format_call<'ast>(ctx: &Context, call: &Call<'ast>, shape: Shape) -> Call
     }
 }
 
+fn is_table_constructor(expression: &Expression) -> bool {
+    match expression {
+        Expression::Value { value, .. } => matches!(**value, Value::TableConstructor(_)),
+        _ => false,
+    }
+}
+
 /// Formats a FunctionArgs node
 pub fn format_function_args<'ast>(
     ctx: &Context,
@@ -295,7 +302,14 @@ pub fn format_function_args<'ast>(
                 }
             }
 
-            if is_multiline {
+            // Handle special case: we want to go multiline, but we have a single argument which is a table constructor
+            // In this case, we want to hug the table braces with the parentheses.
+            // To do this, we format single line, but include the closing parentheses in the shape
+            let hug_table_constructor = is_multiline
+                && arguments.len() == 1
+                && is_table_constructor(arguments.iter().next().unwrap());
+
+            if is_multiline && !hug_table_constructor {
                 // Format start and end brace properly with correct trivia
                 // Calculate to see if the end parentheses requires any additional indentation
                 let end_parens_leading_trivia = vec![create_indent_trivia(ctx, shape)];
@@ -368,8 +382,13 @@ pub fn format_function_args<'ast>(
                 // We don't need to worry about comments here, as if there were comments present, we would have
                 // multiline function args
 
+                // If we are hugging a table constructor with the parentheses, we use a shape increment of 2 to include the closing
+                // parentheses aswell. Otherwise, we just use 1 = opening parentheses.
+                let shape_increment = if hug_table_constructor { 2 } else { 1 };
+
                 let parentheses = format_contained_span(ctx, &parentheses, shape);
-                let arguments = format_punctuated(ctx, arguments, shape + 1, format_expression); // 1 = opening parentheses
+                let arguments =
+                    format_punctuated(ctx, arguments, shape + shape_increment, format_expression);
 
                 FunctionArgs::Parentheses {
                     parentheses,